# swagger.json
{"openapi":"3.0.0","paths":{"/api/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/api/vv1/auth/register/rider":{"post":{"operationId":"AuthController_registerRider_vv1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Rider registered successfully","content":{"application/json":{"schema":{"example":{"user":{"id":"507f1f77bcf86cd799439011","email":"rider@example.com","phone":"+1234567890","role":"RIDER"},"accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"400":{"description":"Validation error"},"409":{"description":"Email or phone already exists"}},"summary":"Register new rider account","tags":["Authentication"]}},"/api/vv1/auth/register/driver":{"post":{"operationId":"AuthController_registerDriver_vv1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Driver registered successfully","content":{"application/json":{"schema":{"example":{"user":{"id":"507f1f77bcf86cd799439012","email":"driver@example.com","phone":"+1234567891","role":"DRIVER"},"accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"400":{"description":"Validation error"},"409":{"description":"Email or phone already exists"}},"summary":"Register new driver account","tags":["Authentication"]}},"/api/vv1/auth/login":{"post":{"operationId":"AuthController_login_vv1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"example":{"user":{"id":"507f1f77bcf86cd799439011","email":"user@example.com","phone":"+1234567890","role":"RIDER"},"accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"400":{"description":"Email or phone is required"},"401":{"description":"Invalid credentials"}},"summary":"Login with email/phone and password","tags":["Authentication"]}},"/api/vv1/auth/refresh":{"post":{"operationId":"AuthController_refresh_vv1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"Tokens refreshed successfully","content":{"application/json":{"schema":{"example":{"accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"401":{"description":"Invalid or expired refresh token"}},"summary":"Refresh access token using refresh token","tags":["Authentication"]}},"/api/vv1/auth/logout":{"post":{"operationId":"AuthController_logout_vv1","parameters":[],"responses":{"200":{"description":"Logged out successfully","content":{"application/json":{"schema":{"example":{"message":"Logged out successfully"}}}}},"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}],"summary":"Logout and invalidate refresh token","tags":["Authentication"]}}},"info":{"title":"Sawari Ride-Hailing API","description":"Production-grade REST API for ride-hailing application. This API provides endpoints for user authentication, ride management, driver operations, payments, and more.","version":"1.0","contact":{}},"tags":[{"name":"health","description":"Health check endpoints"}],"servers":[],"components":{"securitySchemes":{"JWT":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Enter JWT token"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","description":"User email address","example":"rider@example.com"},"phone":{"type":"string","description":"User phone number with country code","example":"+1234567890"},"password":{"type":"string","description":"User password (min 8 chars, must include uppercase, lowercase, number, special char)","example":"SecurePass@123","minLength":8},"role":{"type":"string","description":"User role (RIDER or DRIVER)","enum":["RIDER","DRIVER"],"example":"RIDER"}},"required":["email","phone","password","role"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"User email address (required if phone not provided)","example":"rider@example.com"},"phone":{"type":"string","description":"User phone number (required if email not provided)","example":"+1234567890"},"password":{"type":"string","description":"User password","example":"SecurePass@123"}},"required":["password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string","description":"Refresh token received during login","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"required":["refreshToken"]}}}}